Atualizando com as novas classes e regras de negocio

+++ Domain/Usuarios/Cliente.cs
@@ -18,7 +18,7 @@ namespace Domain.Usuarios
 
         public Pessoa Pessoa { get; private set; }
 
-        public int Numero { get; private set; }
+        public long Numero { get; private set; }
 
     }
 }
